Version 1/backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
google-generativeai
python-multipart
jinja2
python-dotenv